Private pages ids -> tokens, public pages: ids. Added address view page and public token view page
