+++ hooks/tokens.tsx
@@ -25,7 +25,7 @@ export function useTokenList() {
         setTokens(tokenList.tokens)
         setLoading(false)
       } catch (e) {
-        console.error(e.error)
+        console.log(e)
       }
     }
