+++ pages/tokens/mint/[name].tsx
@@ -1,22 +1,18 @@
 import type { NextPage } from 'next'
-import { useRouter } from 'next/router'
-import Link from 'next/link'
-import { useState, useEffect } from 'react'
-import { useForm, UseFormRegister } from 'react-hook-form'
-import WalletLoader from 'components/WalletLoader'
 import InputField from 'components/InputField'
-import { useSigningClient } from 'contexts/cosmwasm'
-import { useToken } from 'hooks/token'
-import { defaultExecuteFee } from 'util/fee'
-import { defaultMemo } from 'util/memo'
-import * as msgs from 'util/messages'
-import * as mt from 'util/types/messages'
+import { useForm } from 'react-hook-form'
+import { useRouter } from 'next/router'
+import { useState } from 'react'
 import { OptionString } from 'util/types/base'
+import { Metadata } from 'util/types/messages'
+import WalletLoader from 'components/WalletLoader'
+import { TokenCard } from 'components/NameCard'
+import { useSigningClient } from 'contexts/cosmwasm'
+import { defaultMintFee, getMintCost } from 'util/fee'
+import { defaultMemo } from 'util/memo'
 import Loader from 'components/Loader'
 import * as R from 'ramda'
 
-// similar to the create form
-// these are the update values
 type FormValues = {
   image: OptionString
   image_data: OptionString
@@ -31,59 +27,35 @@ type FormValues = {
   validator_operator_address: OptionString
 }
 
-const TokenUpdate: NextPage = () => {
-  const router = useRouter()
-  const tokenName = router.query.name as string
+const defaults = {
+  image: null,
+  image_data: null,
+  email: null,
+  external_url: null,
+  public_name: null,
+  public_bio: null,
+  twitter_id: null,
+  discord_id: null,
+  telegram_id: null,
+  keybase_id: null,
+  validator_operator_address: null,
+}
 
+const Mint: NextPage = () => {
+  const router = useRouter()
+  const { signingClient, walletAddress } = useSigningClient()
   const contractAddress = process.env.NEXT_PUBLIC_WHOAMI_ADDRESS as string
+  const [token, setToken] = useState(defaults)
   const [loading, setLoading] = useState(false)
 
-  const [token, setToken] = useState<Metadata>()
-
-  const { signingClient, walletAddress } = useSigningClient()
-
-  useEffect(() => {
-    if (!signingClient) {
-      return
-    }
-
-    const getToken = async () => {
-      //setLoading(true)
-      try {
-        let tokenInfo = await signingClient.queryContractSmart(
-          contractAddress,
-          {
-            nft_info: {
-              token_id: tokenName,
-            },
-          }
-        )
-        setToken(tokenInfo.extension)
-        console.log(tokenInfo)
-        //setLoading(false)
-      } catch (e) {
-        console.log(e)
-      }
-    }
-
-    getToken()
-  }, [tokenName])
-
   const { register, handleSubmit } = useForm<FormValues>({
-    defaultValues: {
-      image: token?.image || null,
-      image_data: token?.image_data || null,
-      email: token?.email || null,
-      external_url: token?.external_url || null,
-      public_name: token?.public_name || null,
-      public_bio: token?.public_bio || null,
-      twitter_id: token?.twitter_id || null,
-      discord_id: token?.discord_id || null,
-      telegram_id: token?.telegram_id || null,
-      keybase_id: token?.keybase_id || null,
-      validator_operator_address: token?.validator_operator_address || null,
-    },
+    defaultValues: defaults,
   })
+
+  if (!router.isReady) {
+    return null
+  }
+  const token_id = router.query.name as string
 
   const onSubmit = async (data: FormValues) => {
     if (!signingClient) {
@@ -107,9 +79,11 @@ const TokenUpdate: NextPage = () => {
     } = data
 
     const msg = {
-      update_metadata: {
-        token_id: tokenName,
-        metadata: {
+      mint: {
+        owner: walletAddress,
+        token_id: token_id,
+        token_uri: null, // TODO - support later
+        extension: {
           image,
           image_data, // TODO - support later
           email,
@@ -126,15 +100,18 @@ const TokenUpdate: NextPage = () => {
     }
 
     try {
-      let updatedToken = await signingClient.execute(
+      const mintCost = getMintCost(token_id)
+
+      let mintedToken = await signingClient.execute(
         walletAddress,
         contractAddress,
         msg,
-        defaultExecuteFee,
-        defaultMemo
+        defaultMintFee,
+        defaultMemo,
+        mintCost
       )
-      if (updatedToken) {
-        router.push(`/ids/${tokenName}`)
+      if (mintedToken) {
+        router.push(`/tokens/${token_id}`)
         setLoading(false)
       }
     } catch (e) {
@@ -165,7 +142,9 @@ const TokenUpdate: NextPage = () => {
         label={i[1] as string}
         register={register}
         optional={i[2] as boolean}
-        onChange={() => {}}
+        onChange={(e) => {
+          setToken((curr) => ({ ...curr, [i[0] as string]: e.target.value }))
+        }}
       />
     ),
     fields
@@ -176,26 +155,28 @@ const TokenUpdate: NextPage = () => {
       {loading ? (
         <Loader />
       ) : (
-        <>
-          <h1 className="text-3xl font-bold">Update your profile</h1>
-
-          <div className="p-6">
-            <p>Update the data associated with your username.</p>
+        <div className="flex flex-wrap">
+          <div className="mr-3">
+            <div className="sticky top-5 mt-5">
+              <TokenCard
+                name={router.query.name as string}
+                token={token as Metadata}
+              />
+            </div>
           </div>
-
           <form onSubmit={handleSubmit(onSubmit)}>
             <>{inputs}</>
 
             <input
               type="submit"
-              className="btn btn-primary btn-lg font-semibold hover:text-base-100 text-2xl w-full"
-              value="Update profile"
+              className="btn btn-primary btn-lg font-semibold hover:text-base-100 text-2xl"
+              value="Create Username"
             />
           </form>
-        </>
+        </div>
       )}
     </WalletLoader>
   )
 }
 
-export default TokenUpdate
+export default Mint
